# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop's customers, mechanics, service tickets, and inventory."
  version: "1.0.0"
host: "127.0.0.1:5000"
basePath: "/"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Enter the word Bearer, a space, and then the token: Bearer <token>"
paths:
  /customers/login:
    post:
      tags:
        - "Customers"
      operationId: "loginCustomer"
      summary: "Log in a customer"
      description: "Validates customer credentials and returns a customer-specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer login credentials."
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful."
          schema:
            $ref: "#/definitions/LoginResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        429:
          $ref: "#/responses/TooManyRequests"
  /customers/:
    post:
      tags:
        - "Customers"
      operationId: "createCustomer"
      summary: "Create a customer"
      description: "Creates a customer. Email addresses must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a customer."
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload"
      responses:
        201:
          description: "Customer created successfully."
          schema:
            $ref: "#/definitions/CustomerResponse"
        400:
          $ref: "#/responses/BadRequest"
        429:
          $ref: "#/responses/TooManyRequests"

    get:
      tags:
        - "Customers"
      operationId: "getCustomers"
      summary: "Get customers"
      description: "Returns a paginated list of customers."
      parameters:
        - in: "query"
          name: "page"
          description: "Page number to retrieve."
          required: false
          type: "integer"
          default: 1
          minimum: 1
        - in: "query"
          name: "per_page"
          description: "Number of customers to return per page."
          required: false
          type: "integer"
          default: 10
          minimum: 1
      responses:
        200:
          description: "Customers retrieved successfully."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CustomerResponse"
        400:
          $ref: "#/responses/BadRequest"
        429:
          $ref: "#/responses/TooManyRequests"

    put:
      tags:
        - "Customers"
      operationId: "updateCustomer"
      summary: "Update the logged-in customer"
      description: "Updates the customer identified by the customer token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Complete replacement values for the customer account."
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload"
      responses:
        200:
          description: "Customer updated successfully."
          schema:
            $ref: "#/definitions/CustomerResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"

    delete:
      tags:
        - "Customers"
      operationId: "deleteCustomer"
      summary: "Delete the logged-in customer"
      description: "Deletes the customer identified by the customer token."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer deleted successfully."
          schema:
            $ref: "#/definitions/MessageResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        429:
          $ref: "#/responses/TooManyRequests"
  /customers/{customer_id}:
    get:
      tags:
        - "Customers"
      operationId: "getCustomer"
      summary: "Get one customer"
      description: "Returns a customer by ID."
      parameters:
        - in: "path"
          name: "customer_id"
          description: "Unique ID of the customer to retrieve."
          required: true
          type: "integer"
          minimum: 1
      responses:
        200:
          description: "Customer retrieved successfully."
          schema:
            $ref: "#/definitions/CustomerResponse"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"
  /mechanics/login:
    post:
      tags:
        - "Mechanics"
      operationId: "loginMechanic"
      summary: "Log in a mechanic"
      description: "Validates mechanic credentials and returns a mechanic-specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic login credentials."
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful."
          schema:
            $ref: "#/definitions/LoginResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        429:
          $ref: "#/responses/TooManyRequests"
  /mechanics/:
    post:
      tags:
        - "Mechanics"
      operationId: "createMechanic"
      summary: "Create a mechanic"
      description: "Creates a mechanic. Email addresses and phone numbers must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a mechanic."
          required: true
          schema:
            $ref: "#/definitions/MechanicPayload"
      responses:
        201:
          description: "Mechanic created successfully."
          schema:
            $ref: "#/definitions/MechanicResponse"
        400:
          $ref: "#/responses/BadRequest"
        429:
          $ref: "#/responses/TooManyRequests"

    get:
      tags:
        - "Mechanics"
      operationId: "getMechanics"
      summary: "Get mechanics"
      description: "Returns all mechanics, or a paginated result when both page and per_page are supplied."
      parameters:
        - in: "query"
          name: "page"
          description: "Page number to retrieve; use together with per_page."
          required: false
          type: "integer"
          minimum: 1
        - in: "query"
          name: "per_page"
          description: "Number of mechanics per page; use together with page."
          required: false
          type: "integer"
          minimum: 1
      responses:
        200:
          description: "Mechanics retrieved successfully."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MechanicResponse"
        429:
          $ref: "#/responses/TooManyRequests"
  /mechanics/experience:
    get:
      tags:
        - "Mechanics"
      operationId: "getMechanicsByExperience"
      summary: "Get mechanics ordered by experience"
      description: "Returns mechanics ordered from most tickets worked to fewest."
      responses:
        200:
          description: "Mechanics retrieved successfully."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MechanicResponse"
        429:
          $ref: "#/responses/TooManyRequests"
  /mechanics/search:
    get:
      tags:
        - "Mechanics"
      operationId: "searchMechanics"
      summary: "Search mechanics by name"
      description: "Returns mechanics whose names contain the supplied text."
      parameters:
        - in: "query"
          name: "name"
          description: "Full or partial mechanic name to search for."
          required: true
          type: "string"
      responses:
        200:
          description: "Mechanics retrieved successfully."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MechanicResponse"
        429:
          $ref: "#/responses/TooManyRequests"
  /mechanics/{mechanic_id}:
    parameters:
      - in: "path"
        name: "mechanic_id"
        description: "Unique ID of the mechanic."
        required: true
        type: "integer"
        minimum: 1

    get:
      tags:
        - "Mechanics"
      operationId: "getMechanic"
      summary: "Get one mechanic"
      description: "Returns a mechanic by ID."
      responses:
        200:
          description: "Mechanic retrieved successfully."
          schema:
            $ref: "#/definitions/MechanicResponse"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"

    put:
      tags:
        - "Mechanics"
      operationId: "updateMechanic"
      summary: "Update a mechanic"
      description: "Updates a mechanic profile. A mechanic may update only their own profile."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Complete replacement values for the mechanic profile."
          required: true
          schema:
            $ref: "#/definitions/MechanicPayload"
      responses:
        200:
          description: "Mechanic updated successfully."
          schema:
            $ref: "#/definitions/MechanicResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        403:
          $ref: "#/responses/Forbidden"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"

    delete:
      tags:
        - "Mechanics"
      operationId: "deleteMechanic"
      summary: "Delete a mechanic"
      description: "Deletes a mechanic profile. A mechanic may delete only their own profile."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Mechanic deleted successfully."
          schema:
            $ref: "#/definitions/MessageResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        403:
          $ref: "#/responses/Forbidden"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"
...
responses:
  BadRequest:
    description: "The request is invalid or authentication is required."
    schema:
      $ref: "#/definitions/ErrorResponse"
  Unauthorized:
    description: "The credentials or token are invalid or expired."
    schema:
      $ref: "#/definitions/ErrorResponse"
  Forbidden:
    description: "The authenticated user is not allowed to perform this operation."
    schema:
      $ref: "#/definitions/ErrorResponse"
  NotFound:
    description: "The requested resource was not found."
    schema:
      $ref: "#/definitions/ErrorResponse"
  TooManyRequests:
    description: "The request exceeded an applicable rate limit."
    schema:
      $ref: "#/definitions/ErrorResponse"
definitions:
  LoginCredentials:
    type: "object"
    required:
      - "email"
      - "password"
    properties:
      email:
        type: "string"
        format: "email"
        example: "user@example.com"
      password:
        type: "string"
        example: "password123"
  LoginResponse:
    type: "object"
    properties:
      auth_token:
        type: "string"
      message:
        type: "string"
        example: "Successfully Logged In"
      status:
        type: "string"
        example: "success"
  CustomerPayload:
    type: "object"
    required:
      - "name"
      - "email"
      - "phone"
      - "password"
    properties:
      name:
        type: "string"
        example: "John Doe"
      email:
        type: "string"
        format: "email"
        example: "john@example.com"
      phone:
        type: "string"
        example: "2222222222"
      password:
        type: "string"
        example: "password123"
  CustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      name:
        type: "string"
        example: "John Doe"
      email:
        type: "string"
        format: "email"
        example: "john@example.com"
      phone:
        type: "string"
        example: "2222222222"
  MechanicPayload:
    type: "object"
    required:
      - "name"
      - "email"
      - "phone"
      - "salary"
      - "password"
    properties:
      name:
        type: "string"
        example: "Jane Mechanic"
      email:
        type: "string"
        format: "email"
        example: "jane@example.com"
      phone:
        type: "string"
        example: "2222222222"
      salary:
        type: "number"
        format: "float"
        example: 75000.0
      password:
        type: "string"
        example: "password123"
  MechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      name:
        type: "string"
        example: "Jane Mechanic"
      email:
        type: "string"
        format: "email"
        example: "jane@example.com"
      phone:
        type: "string"
        example: "2222222222"
      salary:
        type: "number"
        format: "float"
        example: 75000.0
...
  MessageResponse:
    type: "object"
    properties:
      message:
        type: "string"
  ErrorResponse:
    type: "object"
    description: "Error responses may use error or message depending on the handler. Validation errors may include field-specific properties."
    additionalProperties: true
    properties:
      error:
        type: "string"
      message:
        type: "string"